Modify styling of main docs content
This tweaks the CSS used by the main section of the docs pages (not the
sidebar). Styling is largely subjective, but this improves the readability
of the text (heavier weight, more spacing between lines, more contrast
between text and background). It also improves the aesthetics of the inline-
code sections and the buttons at the bottom of each page.

The styles here are inspired by the current nextstrain.org/docs and auspice's
docs (which I really like).