Add custom Nextstrain footer
This modifies the footer to mimic the current Nextstrain (docs) footer.
The content and styles are largely taken from
https://github.com/nextstrain/nextstrain.org/blob/b1e09e57e91ed0c9343e1cd3104877ec3c5344a4/static-site/src/components/Footer/index.jsx

It is not identical -- the RTD footer doesn't fill the available horizontal
space for example -- but it's close.

It's largely written in plain HTML + CSS, as opposed to the source material
which was JSX + Styled Components. There may be a much nicer way to write
this using Jinja functionality.