Sidebar of subprojects provides link back to main docs
This makes it clear that users are within a subproject and provides a simple link to go back to the main docs. This text can be removed / rewritten when we move to a unified TOC between subprojects.

Note that the version does not show up locally, but it does on RTD builds.