Doc: Add a page for the changelog
Switches changelog from Markdown to reStructuredText so it's easier to
include.  (And this is a Sphinx + rST theme after all!)

Switches doc/index.rst from symlinking the README to including it so
that I could add a toctree for Sphinx without impacting the README
displayed on GitHub.  The toctree is necessary now that there's a second
doc page, the changelog.