Show precise versions for stable and development ("latest")
Make it even clearer what software revisions those labels represent.

A `git describe` style string¹ for development ("latest") would be even
better than just the git revision, but that appears to be non-trivial to
implement.

Resolves <https://github.com/nextstrain/augur/issues/817>.

h/t to @corneliusroemer for the suggestion and @huddlej for bringing it
up again in <https://github.com/nextstrain/sphinx-theme/pull/15>.

¹ e.g. "${tag}-${n}-g${sha}" where ${tag} is the last/closest tag to the
  current commit, ${n} is the number of commits since that ${tag}, and
  ${sha} is the shortened id of the current commit.