Dock the floating RTD versions panel to the sidebar
…by lying to RTD about what theme we really are. 🙈

Resolves a long-standing issue with the controls for the versions panel
being at the bottom left of the page (in the sidebar) but the panel
content floating as a "badge" in the lower right corner of the page.
See <https://github.com/nextstrain/docs.nextstrain.org/issues/76> for
the diagnosis.

While fixing the underlying issue with RTD's code would be best, it
isn't clear to me how to go about doing so, e.g. how to introspect theme
inheritance.