Include sphinx-copybutton with nextstrain-sphinx-theme
This extension has been recently used in a few of Nextstrain's doc projects:

- https://github.com/nextstrain/docs.nextstrain.org/commit/cc2e2a85c66e36591d83b34b7195b1eff18b3ce7
- https://github.com/nextstrain/ncov/commit/650820252030f23267dd6ba8f5d207ff3cbf466a

Adding it to this package will make Nextstrain's doc projects more consistent going forwards.

Note: for this to work properly, this package must be used as an extension in addition to usage as a theme.
See eabffdf as an example.